+++ crates/symbolicator/src/endpoints/applecrashreport.rs
@@ -23,6 +23,7 @@ pub async fn handle_apple_crash_report_request(
     let mut sources = service.config().default_sources();
     let mut scraping = Default::default();
     let mut options = RequestOptions::default();
+    let mut platform = None;
 
     while let Some(field) = multipart.next_field().await? {
         match field.name() {
@@ -43,14 +44,24 @@ pub async fn handle_apple_crash_report_request(
                 let data = read_multipart_data(field, 1024 * 1024).await?; // 1Mb
                 options = serde_json::from_slice(&data)?
             }
+            Some("platform") => {
+                let data = read_multipart_data(field, 1024 * 1024).await?; // 1Mb
+                platform = serde_json::from_slice(&data)?
+            }
             _ => (), // Always ignore unknown fields.
         }
     }
 
     let report = report.ok_or((StatusCode::BAD_REQUEST, "missing apple crash report"))?;
 
-    let request_id =
-        service.process_apple_crash_report(params.scope, report, sources, scraping, options)?;
+    let request_id = service.process_apple_crash_report(
+        platform,
+        params.scope,
+        report,
+        sources,
+        scraping,
+        options,
+    )?;
 
     match service.get_response(request_id, params.timeout).await {
         Some(response) => Ok(Json(response)),